# api/openapi.yaml
openapi: 3.0.3
info:
  title: Amplify
  version: 0.0.1
  description: Amplify public API
servers:
  - url: /api
paths:
...
  /v0/queue:
    get:
      summary: Amplify work queue
      description: Lists all workflows and jobs in the queue
      parameters:
        - in: query
          name: createdBefore
          schema:
            type: string
            format: date-time
          description: Filter for items created before this date-time
        - in: query
          name: createdAfter
          schema:
            type: string
            format: date-time
          description: Filter for items created after this date-time
        - in: query
          name: limit
          schema:
            type: integer
          description: The numbers of items to return
      responses:
        "200":
          description: Ok.
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/queue"
            text/html:
              schema:
                type: string
                example: <html><body>Body text</body></html>
    post:
      summary: Run all workflows for a CID (not recommended)
      description: This endpoint is for form support only, it is not recommended for normal use because it is not imdepotent. See PUT /api/v0/queue/{id}
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/executionRequest"
      responses:
        "202":
          description: Accepted.
        "400":
          description: Error submitting request.
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/errors"
            text/html:
              schema:
                type: string
                example: <html><body>Body text</body></html>
...
components:
  schemas:
...
    executionRequest:
      type: object
      required: ["cid"]
      properties:
        cid:
          type: string
          example: bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi
    itemMetadata:
      type: object
      required: ["submitted", "status"]
      properties:
        submitted:
          type: string
          example: "2019-10-12T07:20:50.52Z"
        started:
          type: string
          example: "2019-10-12T07:20:50.52Z"
        ended:
          type: string
          example: "2019-10-12T07:20:50.52Z"
        status:
          type: string
          example: "running"
    item:
      type: object
      required: ["id", "type", "metadata"]
      properties:
        id:
          type: string
          example: 740a5879-429c-4aa2-9807-eb200b1ae754
        type:
          type: string
          example: item
        links:
          $ref: "#/components/schemas/links"
        metadata:
          $ref: "#/components/schemas/itemMetadata"
...
    queue:
      type: object
      properties:
        meta:
          $ref: "#/components/schemas/pageMeta"
        data:
          type: array
          items:
            $ref: "#/components/schemas/item"
        links:
          $ref: "#/components/schemas/links"
...
    pageMeta:
      type: object
      properties:
        totalPages:
          type: integer
          example: 13
          description: Total number of pages in paginated result.
    links:
      type: object
    errors:
      type: array
      items:
        $ref: "#/components/schemas/error"
    error:
      type: object
      properties:
        title:
          type: string
          description: A short, human-readable summary of the problem that SHOULD NOT change from occurrence to occurrence of the problem, except for purposes of localization.
          example: Job not found
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
          example: Could not find a reference to a job called "metadata"