# pkg/api/api.go
package api

import (
	"bytes"
	"context"
	"embed"
	"encoding/json"
	"errors"
	"fmt"
	"html/template"
	"net/http"
	"sync"
	"time"

	"github.com/bacalhau-project/amplify/pkg/dag"
	"github.com/bacalhau-project/amplify/pkg/db"
	"github.com/bacalhau-project/amplify/pkg/item"
	"github.com/bacalhau-project/amplify/pkg/queue"
	"github.com/bacalhau-project/amplify/pkg/task"
	"github.com/bacalhau-project/amplify/pkg/util"
	openapi_types "github.com/deepmap/oapi-codegen/pkg/types"
	"github.com/google/uuid"
	"github.com/rs/zerolog/log"
)

var (
	funcs = template.FuncMap{"marshall": func(v interface{}) string {
		b, _ := json.Marshal(v)
		return string(b)
	}}
)

// content holds our static web server content.
//
//go:embed templates/*
var content embed.FS

type amplifyAPI struct {
	*sync.Mutex
	er   item.QueueRepository
	tf   task.TaskFactory
	tmpl *template.Template
}

func NewAmplifyAPI(er item.QueueRepository, tf task.TaskFactory) (*amplifyAPI, error) {
	tmpl := template.New("master").Funcs(funcs)
	tmpl, err := tmpl.ParseFS(content, "templates/*.html.tmpl")
	if err != nil {
		return nil, err
	}
	return &amplifyAPI{
		er:   er,
		tf:   tf,
		tmpl: tmpl,
	}, nil
}

// Amplify home
// (GET /v0)
func (a *amplifyAPI) GetV0(w http.ResponseWriter, r *http.Request) {
	log.Ctx(r.Context()).Trace().Msg("GetV0")
	home := &Home{
		Type: util.StrP("home"),
		Links: util.MapP(map[string]interface{}{
			"self":  "/api/v0",
			"queue": "/api/v0/queue",
			"jobs":  "/api/v0/jobs",
			"graph": "/api/v0/graph",
		}),
	}
	switch r.Header.Get("Accept") {
	case "application/json":
		fallthrough
	case "application/vnd.api+json":
		w.Header().Set("Content-Type", "application/vnd.api+json")
		w.Header().Set("Access-Control-Allow-Origin", "*")
		err := json.NewEncoder(w).Encode(home)
		if err != nil {
			sendError(r.Context(), w, http.StatusInternalServerError, "Could not render JSON", err.Error())
			return
		}
	default:
		err := a.writeHTML(w, "home.html.tmpl", home)
		if err != nil {
			sendError(r.Context(), w, http.StatusInternalServerError, "Could not render HTML", err.Error())
			return
		}
	}
}

// List all Amplify jobs
// (GET /v0/jobs)
func (a *amplifyAPI) GetV0Jobs(w http.ResponseWriter, r *http.Request) {
	log.Ctx(r.Context()).Trace().Msg("GetV0Jobs")
	jobs, err := a.getJobs()
	if err != nil {
		sendError(r.Context(), w, http.StatusInternalServerError, "Could not get jobs", err.Error())
		return
	}
	switch r.Header.Get("Accept") {
	case "application/json":
		fallthrough
	case "application/vnd.api+json":
		w.Header().Set("Content-Type", "application/vnd.api+json")
		w.Header().Set("Access-Control-Allow-Origin", "*")
		err := json.NewEncoder(w).Encode(jobs)
		if err != nil {
			sendError(r.Context(), w, http.StatusInternalServerError, "Could not render JSON", err.Error())
			return
		}
	default:
		err := a.writeHTML(w, "jobs.html.tmpl", jobs)
		if err != nil {
			sendError(r.Context(), w, http.StatusInternalServerError, "Could not render HTML", err.Error())
			return
		}
	}
}

// Get a job by id
// (GET /v0/jobs/{id})
func (a *amplifyAPI) GetV0JobsId(w http.ResponseWriter, r *http.Request, id string) {
	log.Ctx(r.Context()).Trace().Str("id", id).Msg("GetV0JobsId")
	j, err := a.getJob(id)
	if err != nil {
		if errors.Is(err, task.ErrJobNotFound) {
			sendError(r.Context(), w, http.StatusNotFound, "Job not found", fmt.Sprintf("Job %s not found", id))
		} else {
			sendError(r.Context(), w, http.StatusInternalServerError, "Could not get job", err.Error())
		}
		return
	}
	switch r.Header.Get("Accept") {
	case "application/json":
		fallthrough
	case "application/vnd.api+json":
		w.Header().Set("Content-Type", "application/vnd.api+json")
		w.Header().Set("Access-Control-Allow-Origin", "*")
		err := json.NewEncoder(w).Encode(j)
		if err != nil {
			sendError(r.Context(), w, http.StatusInternalServerError, "Could not render JSON", err.Error())
			return
		}
	default:
		err := a.writeHTML(w, "job.html.tmpl", j)
		if err != nil {
			sendError(r.Context(), w, http.StatusInternalServerError, "Could not render HTML", err.Error())
			return
		}
	}
}

// Amplify work queue
// (GET /v0/queue)
func (a *amplifyAPI) GetV0Queue(w http.ResponseWriter, r *http.Request, params GetV0QueueParams) {
	log.Ctx(r.Context()).Trace().Msg("GetV0Queue")
	paginationParams := item.PaginationParams{
		Limit:         10,
		CreatedAfter:  time.Time{},
		CreatedBefore: time.Now(),
	}
	if params.Limit != nil {
		paginationParams.Limit = *params.Limit
	}
	if params.CreatedBefore != nil {
		paginationParams.CreatedBefore = *params.CreatedBefore
	}
	if params.CreatedAfter != nil {
		paginationParams.CreatedAfter = *params.CreatedAfter
	}
	executions, err := a.getQueue(r.Context(), paginationParams)
	if err != nil {
		sendError(r.Context(), w, http.StatusInternalServerError, "Could not get executions", err.Error())
		return
	}
	switch r.Header.Get("Accept") {
	case "application/json":
		fallthrough
	case "application/vnd.api+json":
		w.Header().Set("Content-Type", "application/vnd.api+json")
		w.Header().Set("Access-Control-Allow-Origin", "*")
		err := json.NewEncoder(w).Encode(executions)
		if err != nil {
			sendError(r.Context(), w, http.StatusInternalServerError, "Could not render JSON", err.Error())
			return
		}
	default:
		err := a.writeHTML(w, "queue.html.tmpl", executions)
		if err != nil {
			sendError(r.Context(), w, http.StatusInternalServerError, "Could not render HTML", err.Error())
			return
		}
	}
}

// Get an item from the queue by id
// (GET /v0/queue/{id})
func (a *amplifyAPI) GetV0QueueId(w http.ResponseWriter, r *http.Request, id openapi_types.UUID) {
	log.Ctx(r.Context()).Trace().Str("id", id.String()).Msg("GetV0QueueId")
	e, err := a.getItemDetail(r.Context(), id)
	if err != nil {
		if errors.Is(err, db.ErrNotFound) {
			sendError(r.Context(), w, http.StatusNotFound, "Execution not found", fmt.Sprintf("Execution %s not found", id.String()))
		} else {
			sendError(r.Context(), w, http.StatusInternalServerError, "Could not get execution", err.Error())
		}
		return
	}
	switch r.Header.Get("Accept") {
	case "application/json":
		fallthrough
	case "application/vnd.api+json":
		w.Header().Set("Content-Type", "application/vnd.api+json")
		w.Header().Set("Access-Control-Allow-Origin", "*")
		err := json.NewEncoder(w).Encode(e)
		if err != nil {
			sendError(r.Context(), w, http.StatusInternalServerError, "Could not render JSON", err.Error())
			return
		}
	default:
		err := a.writeHTML(w, "queueItem.html.tmpl", e)
		if err != nil {
			sendError(r.Context(), w, http.StatusInternalServerError, "Could not render HTML", err.Error())
			return
		}
	}
}

// Run all workflows for a CID
// (PUT /v0/queue/{id})
func (a *amplifyAPI) PutV0QueueId(w http.ResponseWriter, r *http.Request, id openapi_types.UUID) {
	log.Ctx(r.Context()).Trace().Str("id", id.String()).Msg("PutV0QueueId")
	w.Header().Set("Access-Control-Allow-Origin", "*")
	var body ExecutionRequest
	switch r.Header.Get("Content-Type") {
	case "application/json":
		fallthrough
	case "application/vnd.api+json":
		// Parse request body
		err := json.NewDecoder(r.Body).Decode(&body)
		if err != nil {
			sendError(r.Context(), w, http.StatusBadRequest, "Could not parse request body", err.Error())
			return
		}
	default:
		sendError(r.Context(), w, http.StatusBadRequest, "Wrong content type", "The Content-Type header is not set according to the API spec.")
		return
	}

	a.createExecution(r.Context(), w, id, body.Cid)
}

// Run all workflows for a CID (not recommended)
// (POST /api/v0/queue)
func (a *amplifyAPI) PostV0Queue(w http.ResponseWriter, r *http.Request) {
	log.Ctx(r.Context()).Trace().Msg("PostV0Queue")
	var body ExecutionRequest
	w.Header().Set("Access-Control-Allow-Origin", "*")
	switch r.Header.Get("Content-Type") {
	case "application/x-www-form-urlencoded":
		err := r.ParseForm()
		if err != nil {
			sendError(r.Context(), w, http.StatusBadRequest, "Could not parse request body", err.Error())
			return
		}
		body.Cid = r.FormValue("cid")
	default:
		sendError(r.Context(), w, http.StatusBadRequest, "Wrong content type", "The Content-Type header is not set according to the API spec.")
		return
	}
	a.createExecution(r.Context(), w, uuid.New(), body.Cid)
}

func (a *amplifyAPI) createExecution(ctx context.Context, w http.ResponseWriter, executionID uuid.UUID, cid string) {
	err := a.CreateExecution(ctx, executionID, cid)
	if err != nil {
		if err == queue.ErrQueueFull {
			sendError(ctx, w, http.StatusTooManyRequests, "Queue full", err.Error())
			return
		} else {
			sendError(ctx, w, http.StatusInternalServerError, "Could not create execution", err.Error())
			return
		}
	}
	w.WriteHeader(202)
}

func (a *amplifyAPI) CreateExecution(ctx context.Context, executionID uuid.UUID, cid string) error {
	return a.er.Create(ctx, item.ItemParams{
		ID:  executionID,
		CID: cid,
	})
}

// Get Amplify work graph
// (GET /v0/graph)
func (a *amplifyAPI) GetV0Graph(w http.ResponseWriter, r *http.Request) {
	log.Ctx(r.Context()).Trace().Msg("GetV0Workflows")
	nn := a.tf.NodeNames()
	graph := make([]NodeConfig, len(nn))
	for idx, n := range nn {
		node, err := a.tf.GetNode(n)
		if err != nil {
			sendError(r.Context(), w, http.StatusInternalServerError, "Could not get node", err.Error())
			return
		}
		inputs := make([]NodeInput, len(node.Inputs))
		for i, input := range node.Inputs {
			inputs[i] = NodeInput{
				OutputId:  &input.OutputID,
				Path:      &input.Path,
				Root:      &input.Root,
				StepId:    &input.NodeID,
				Predicate: &input.Predicate,
			}
		}
		outputs := make([]NodeOutput, len(node.Outputs))
		for i, output := range node.Outputs {
			outputs[i] = NodeOutput{
				Id:   &output.ID,
				Path: &output.Path,
			}
		}
		graph[idx] = NodeConfig{
			Id:      &node.ID,
			Inputs:  &inputs,
			JobId:   &node.JobID,
			Outputs: &outputs,
		}
	}
	outputGraph := Graph{
		Data: &graph,
		Links: &Links{
			"self": "/api/v0/graph",
			"home": "/api/v0",
		},
	}
	switch r.Header.Get("Accept") {
	case "application/json":
		fallthrough
	case "application/vnd.api+json":
		w.Header().Set("Content-Type", "application/vnd.api+json")
		w.Header().Set("Access-Control-Allow-Origin", "*")
		err := json.NewEncoder(w).Encode(outputGraph)
		if err != nil {
			sendError(r.Context(), w, http.StatusInternalServerError, "Could not render JSON", err.Error())
			return
		}
	default:
		err := a.writeHTML(w, "graph.html.tmpl", outputGraph)
		if err != nil {
			sendError(r.Context(), w, http.StatusInternalServerError, "Could not render HTML", err.Error())
			return
		}
	}
}

func (a *amplifyAPI) getJobs() (*Jobs, error) {
	jobList := make([]Job, len(a.tf.JobNames()))
	for i, id := range a.tf.JobNames() {
		j, err := a.getJob(id)
		if err != nil {
			return nil, err
		}
		jobList[i] = *j
	}
	return &Jobs{
		Data: &jobList,
		Links: &Links{
			"self": "/api/v0/jobs",
			"home": "/api/v0",
		},
	}, nil
}

func (a *amplifyAPI) getJob(jobId string) (*Job, error) {
	j, err := a.tf.GetJob(jobId)
	if err != nil {
		return nil, err
	}
	return &Job{
		Type:       "job",
		Image:      j.Image,
		Entrypoint: &j.Entrypoint,
		Id:         j.ID,
		Links: &Links{
			"self": fmt.Sprintf("/api/v0/jobs/%s", j.ID),
			"list": "/api/v0/jobs",
			"home": "/api/v0",
		},
	}, nil
}

func (a *amplifyAPI) getQueue(ctx context.Context, params item.PaginationParams) (*Queue, error) {
	e, err := a.er.List(ctx, params)
	if err != nil {
		return nil, err
	}
	items := make([]Item, len(e))
	for i, item := range e {
		items[i] = *buildItem(ctx, item)
	}
	lastItems, err := a.er.List(ctx, item.PaginationParams{
		Limit:         params.Limit,
		CreatedBefore: time.Now(),
		CreatedAfter:  time.Time{},
		Reverse:       true,
	})
	if err != nil {
		return nil, err
	}
	lastItemTime := time.Now()
	if len(lastItems) > 0 {
		lastItemTime = lastItems[len(lastItems)-1].Metadata.CreatedAt.Add(1 * time.Second)
	}
	q := &Queue{
		Data: &items,
		Links: &Links{
			"self":  fmt.Sprintf("/api/v0/queue?limit=%d&createdAfter=%s&createdBefore=%s", params.Limit, params.CreatedAfter.UTC().Format(time.RFC3339), params.CreatedBefore.UTC().Format(time.RFC3339)),
			"home":  "/api/v0",
			"first": fmt.Sprintf("/api/v0/queue?limit=%d", params.Limit),
			"last":  fmt.Sprintf("/api/v0/queue?limit=%d&createdBefore=%s", params.Limit, lastItemTime.Format(time.RFC3339)),
		},
	}
	if len(items) > 0 {
		lastTime := items[len(items)-1].Metadata.Submitted
		(*q.Links)["next"] = fmt.Sprintf("/api/v0/queue?limit=%d&createdBefore=%s", params.Limit, lastTime)
	}
	return q, nil
}

func makeNode(ctx context.Context, dagNode dag.Node[dag.IOSpec], rootId openapi_types.UUID) Node {
	child, err := dagNode.Get(ctx)
	if err != nil {
		log.Ctx(ctx).Error().Err(err).Msg("Could not get node")
		return Node{}
	}
	inputs := make([]ExecutionRequest, len(child.Inputs))
	for idx, input := range child.Inputs {
		inputs[idx] = ExecutionRequest{
			Cid: input.CID(),
		}
	}
	outputs := make([]ExecutionRequest, len(child.Outputs))
	for idx, output := range child.Outputs {
		outputs[idx] = ExecutionRequest{
			Cid: output.CID(),
		}
	}
	node := Node{
		Id:      rootId,
		Name:    util.StrP(child.Name),
		Inputs:  inputs,
		Outputs: outputs,
		Result: &ItemResult{
			Id:      util.StrP(child.Results.ID),
			Stderr:  util.StrP(child.Results.StdErr),
			Stdout:  util.StrP(child.Results.StdOut),
			Skipped: util.BoolP(child.Results.Skipped),
		},
		Links: &Links{
			"self": fmt.Sprintf("/api/v0/queue/%s", rootId),
			"list": "/api/v0/queue",
		},
		Metadata: ItemMetadata{
			Submitted: child.Metadata.CreatedAt.Format(time.RFC3339),
			Status:    child.Metadata.Status,
		},
	}
	if !child.Metadata.StartedAt.IsZero() {
		node.Metadata.Started = util.StrP(child.Metadata.StartedAt.Format(time.RFC3339))
	}
	if !child.Metadata.EndedAt.IsZero() {
		node.Metadata.Ended = util.StrP(child.Metadata.EndedAt.Format(time.RFC3339))
	}
	if len(child.Children) > 0 {
		children := make([]Node, len(child.Children))
		for idx, c := range child.Children {
			children[idx] = makeNode(ctx, c, rootId)
		}
		node.Children = &children
	}
	return node
}

func (a *amplifyAPI) getItemDetail(ctx context.Context, executionId openapi_types.UUID) (*Node, error) {
	i, err := a.er.Get(ctx, executionId)
	if err != nil {
		return nil, err
	}
	dag := make([]Node, len(i.RootNodes))
	for idx, child := range i.RootNodes {
		dag[idx] = makeNode(ctx, child, executionId)
	}
	results := GetLeafOutputs(ctx, i.RootNodes)
	results = dedup(results)
	outputs := make([]ExecutionRequest, len(results))
	for idx, output := range results {
		outputs[idx] = ExecutionRequest{
			Cid: output,
		}
	}

	v := &Node{
		Id:   i.ID,
		Type: "node",
		Inputs: []ExecutionRequest{{
			Cid: i.CID,
		}},
		Outputs: outputs,
		Metadata: ItemMetadata{
			Submitted: i.Metadata.CreatedAt.Format(time.RFC3339),
			Status:    childStatusesToList(ctx, i.RootNodes),
		},
		Links: &Links{
			"self": fmt.Sprintf("/api/v0/queue/%s", i.ID),
			"list": "/api/v0/queue",
		},
		Children: &dag,
	}
	if !i.Metadata.StartedAt.IsZero() {
		v.Metadata.Started = util.StrP(i.Metadata.StartedAt.Format(time.RFC3339))
	}
	if !i.Metadata.EndedAt.IsZero() {
		v.Metadata.Ended = util.StrP(i.Metadata.EndedAt.Format(time.RFC3339))
	}
	return v, nil
}

func parseChildrenStatus(ctx context.Context, children []dag.Node[dag.IOSpec], statuses map[int32]string) map[int32]string {
	for _, child := range children {
		_, ok := statuses[child.ID()]
		if ok {
			continue
		}
		c, err := child.Get(ctx)
		if err != nil {
			log.Ctx(ctx).Error().Err(err).Msg("Could not get node")
			continue
		}
		statuses[child.ID()] = c.Metadata.Status
		if len(c.Children) > 0 {
			statuses = parseChildrenStatus(ctx, c.Children, statuses)
		}
	}
	return statuses
}

func childStatusesToList(ctx context.Context, children []dag.Node[dag.IOSpec]) string {
	statuses := parseChildrenStatus(ctx, children, make(map[int32]string, len(children)))
	jobsCompleted := 0
	for _, v := range statuses {
		if v == dag.Finished.String() {
			jobsCompleted++
		}
	}
	return fmt.Sprintf("%.0f%%", 100*float32(jobsCompleted)/float32(len(statuses)))
}

func buildItem(ctx context.Context, i *item.Item) *Item {
	v := Item{
		Id:   i.ID.String(),
		Type: "item",
		Metadata: ItemMetadata{
			Submitted: i.Metadata.CreatedAt.Format(time.RFC3339),
			Started:   util.StrP(i.Metadata.StartedAt.Format(time.RFC3339)),
			Ended:     util.StrP(i.Metadata.EndedAt.Format(time.RFC3339)),
			Status:    childStatusesToList(ctx, i.RootNodes),
		},
		Links: &Links{
			"self": fmt.Sprintf("/api/v0/queue/%s", i.ID),
			"list": "/api/v0/queue",
		},
	}
	if !i.Metadata.StartedAt.IsZero() {
		v.Metadata.Started = util.StrP(i.Metadata.StartedAt.Format(time.RFC3339))
	}
	if !i.Metadata.EndedAt.IsZero() {
		v.Metadata.Ended = util.StrP(i.Metadata.EndedAt.Format(time.RFC3339))
	}
	return &v
}

func (a *amplifyAPI) writeHTML(w http.ResponseWriter, templateName string, data interface{}) error {
	t := a.tmpl.Lookup(templateName).Option("missingkey=error")
	buf := &bytes.Buffer{}
	err := t.Execute(buf, data)
	if err != nil {
		return err
	} else {
		w.Header().Set("Content-Type", "text/html")
		_, err := buf.WriteTo(w)
		if err != nil {
			return err
		}
	}
	return nil
}

func sendError(ctx context.Context, w http.ResponseWriter, statusCode int, userErr, devErr string) {
	log.Ctx(ctx).Warn().Int("status", statusCode).Str("user", userErr).Str("dev", devErr).Msg("API Error")
	e := Error{
		Title:  &userErr,
		Detail: &devErr,
	}
	w.Header().Set("Content-Type", "application/vnd.api+json")
	w.Header().Set("Access-Control-Allow-Origin", "*")
	w.WriteHeader(statusCode)
	err := json.NewEncoder(w).Encode(&Errors{e})
	if err != nil {
		log.Ctx(ctx).Error().Err(err).Msg("Error sending error response")
	}
}

// getLeafOutputs returns the outputs of the leaf nodes in the DAG
func GetLeafOutputs(ctx context.Context, dag []dag.Node[dag.IOSpec]) []string {
	var results []string
	for _, child := range dag {
		c, err := child.Get(ctx)
		if err != nil {
			log.Ctx(ctx).Error().Err(err).Msg("Could not get node")
			return nil
		}
		if len(c.Children) == 0 {
			for _, output := range c.Outputs {
				results = append(results, output.CID())
			}
		} else {
			results = append(results, GetLeafOutputs(ctx, c.Children)...)
		}
	}
	return results
}

// dedup removes duplicate strings from a slice
func dedup(s []string) []string {
	m := make(map[string]bool)
	for _, v := range s {
		m[v] = true
	}
	var results []string
	for k := range m {
		results = append(results, k)
	}
	return results
}